Decouple CLI commands and start decoupling and cleaning up consolidator.ts
